old/requirements.txt
openai>=0.27.0
faiss-cpu>=1.7.3
numpy>=1.21.0